requirements.txt
pipecat-ai[google,silero,webrtc]>=0.0.102,<1.0
fastapi>=0.115.0,<1.0
uvicorn>=0.32.0,<1.0
python-dotenv>=1.0.0,<2.0
aiohttp>=3.9.0,<4.0
aiosqlite>=0.20.0,<1.0
python-dateutil>=2.9.0,<3.0
openai>=1.0.0,<2.0
boto3>=1.35.0,<2.0
openpyxl>=3.1.0,<4.0
google-auth>=2.0.0,<3.0
google-api-python-client>=2.0.0,<3.0
